refactor: 유저 정보를 메인 페이지에서 불러와서 redux에 저장하도록 변경
